requirements.txt
Django==2.2.12
GitPython==3.1.2
requests==2.23.0